refactor(model): extract Telemetry to dedicated file
- Move Telemetry to modules/model/telemetry.go for reuse and clarity. No behavior change.
